Removed old dispatch processing code
